Support configs for applets
Applets have the capability of defining configuration which is acts as
user preferences. The preferences are chosen by the user via a UI, and
provided to the Applet through pixlet.

This patch adds support for providing the user preferences as a part
of config.json, which is then passed on to the applet.